fix(diagnostics): move connection diagnostics from dead code to actual entry point
The diagnostics setup was in server.py's `if __name__ == "__main__":` block,
but that code never executes because server.py is imported by __main__.py,
not run directly.

Moved all diagnostics initialization to __main__.py where it actually runs:
- Signal handlers for graceful shutdown
- Connection start/end logging
- Heartbeat monitor (30s interval)
- Cleanup handlers with diagnostic summaries
- Error pattern detection

Also added get_unified_collector() export helper in server.py for cleaner
imports in __main__.py.

This ensures connection diagnostics actually work instead of being silently
skipped.
